chore(autoscaling-common): drop stability entry
Motivation

This module contains a set of utility methods and classes to be reused
across other construct libraries. Customers will not be using these
libraries directly and their utility will be directly governed by the
needs of the consuming packages. Stating a 'stability' level similar to
real construct libraries used by customers is misleading.

Further, this module is consumed by aws-applicationautoscaling and
aws-autoscaling. Both of these construct libraries are already stable.
So it makes little sense for this to be marked 'experimental'.